==> notebooks/1-getting_data.ipynb <==
import pandas as pd
import sys

%load_ext autoreload
%autoreload 2

sys.path.append('../src')

from cleaning import droping_columns, sort_by_elements, re_label, new_index, rename_columns, order

# Data

#### First dataset

!kaggle datasets download -d eward96/skincare-products-and-their-ingredients

!tar -xzvf skincare-products-and-their-ingredients.zip

!mv ../notebooks/skincare_products.csv ../data/skincare_products.csv

!rm skincare-products-and-their-ingredients.zip

first = pd .read_csv('../data/skincare_products.csv')

#### Second dataset 

!kaggle datasets download -d raghadalharbi/all-products-available-on-sephora-website

!tar -xzvf all-products-available-on-sephora-website.zip

!mv ../notebooks/sephora_website_dataset.csv ../data/sephora_website_dataset.csv

!rm all-products-available-on-sephora-website.zip

second = pd.read_csv('../data/sephora_website_dataset.csv')

#### Third dataset 

third = pd.read_csv('../data/AB1.csv', usecols=['brand', 'product_name', 'ingredients'])

#### Fourth dataset 

fourth = pd.read_csv('../data/AB2.csv', usecols=['brand', 'product_name', 'ingredients'])

#### Fifth dataset 

fifth = pd.read_csv('../data/Prod1.csv', usecols=['brand', 'product_name', 'ingredients'])

#### Sixth dataset 

sixth = pd.read_csv('../data/Prod2.csv', usecols=['brand', 'product_name', 'ingredients'])

#### Seventh dataset 

seventh = pd.read_csv('../data/red1.csv', usecols=['brand', 'product_name', 'ingredients'])

# Cleaning

#### Second dataset 

I prioritize the second dataset since it is the most complete

# Droping some columns I won't use

to_drop_second = ['id', 'size', 'number_of_reviews', 'love', 'value_price', 'MarketingFlags', 'MarketingFlags_content', 'options', 'details', 'how_to_use', 'online_only', 'exclusive', 'limited_edition', 'limited_time_offer']

second = droping_columns(second, to_drop_second)

# Among all the categories I just want some

to_stay = ['Face Serums', 'Moisturizers', 'Eye Creams & Treatments', 'Face Sunscreen', 'Face Masks', 'Face Wash & Cleansers', 'Decollete & Neck Creams', 'Face Oils', 'Tinted Moisturizer', 'Facial Peels', 'Exfoliators', 'BB & CC Cream', 'Toners', 'Sheet Masks', 'Face Wash', 'Skincare Sets', 'Moisturizer & Treatments', 'Eye Cream', 'Sunscreen', 'Face Sets', 'Anti-Aging', 'Facial Rollers', 'Eye Masks', 'Face Wipes', 'Night Creams', 'Blemish & Acne Treatments', 'Scrub & Exfoliants', 'For Face', 'Facial Cleansing Brushes', 'BB & CC Creams', 'After Sun Care']

second = sort_by_elements(second, second.category, to_stay)

# Relabeling the categories due to some are repetead

allowed = {'After Sun Care':'Sunscreen & After Sun Care',
     'Anti-Aging':'Anti-Aging',
     'BB & CC Cream':'BB & CC Cream',
     'BB & CC Creams':'BB & CC Cream',
     'Blemish & Acne Treatments':'Blemish & Acne Treatment',
     'Decollete & Neck Creams':'Decollete & Neck Cream',
     'Exfoliators':'Exfoliator',
     'Eye Cream':'Eye Cream',
     'Eye Creams & Treatments':'Eye Cream',
     'Eye Masks':'Eye Mask',
     'Face Masks':'Face Mask',
     'Face Oils':'Face Oil',
     'Face Serums':'Face Serum',
     'Face Sets':'Face Set',
     'Face Sunscreen':'Sunscreen & After Sun Care',
     'Face Wash':'Cleanser',
     'Face Wash & Cleansers':'Cleanser',
     'Face Wipes':'Face Wipe',
     'Facial Cleansing Brushes':'Facial Cleansing Brush',
     'Facial Peels':'Facial Peel',
     'Facial Rollers':'Facial Roller',
     'For Face':'Face Tanner',
     'Moisturizer & Treatments':'Moisturizer',
     'Moisturizers':'Moisturizer',
     'Night Creams':'Night Cream',
     'Scrub & Exfoliants':'Exfoliator',
     'Sheet Masks':'Face Mask',
     'Skincare Sets':'Face Set',
     'Sunscreen':'Sunscreen & After Sun Care',
     'Tinted Moisturizer':'Moisturizer',
     'Toners':'Toner',
          }

second = re_label(second, allowed)

# Reseting the index

second = new_index(second)

second

# From dollar to euros

second.price = second.price.apply(lambda x: round(number=x*0.83, ndigits=2))

# Rename columns

names = {'brand':'Brand',
         'category':'Category',
         'name': 'Name',
         'rating':'Rating',
         'price':'Price',
         'URL':'URL',
         'ingredients':'Ingredients'
}

second = rename_columns(second,names)

# New order of columns

columns = ['Brand', 'Category', 'Name', 'Ingredients', 'URL', 'Price', 'Rating']

second = order(second, columns)

second

# Export

second.to_csv('../data/clean_data.csv')
