==> notebooks/3-adding_new_data_beautiful_soup.ipynb <==
import pandas as pd

import requests

from bs4 import BeautifulSoup

import sys

 %load_ext autoreload
%autoreload 2

sys.path.append('../src')

from bs_work import from_serie_to_list, skin_type_category

# Previous dataset

data = pd.read_csv('../data/clean_data.csv', index_col='Unnamed: 0')

I divied the dataset so it is easy to find with BS

data_1 = data[0:279]

data_2 = data[279:558]

data_3 = data[558:837]

data_4 = data[837:1116]

data_5 = data[1116:1395]

data_6 = data[1395:1674]

data_7 = data[1674:1953]

data_8 = data[1953:2232]

data_9 = data[2232:2511]

# links_1 = []
# for d in data_1.URL:
#     links_1.append(d)

# links_2 = []
# for d in data_2.URL:
#     links_2.append(d)

# Adding data 

Using beutiful soup to add info to the previous dataframe

One example:

url = 'https://www.sephora.com/product/clarifying-cleanser-P429954?icid2=products grid:p429954'

response = requests.get(url)

response.status_code

soup = BeautifulSoup(response.content, 'html.parser')

skin_type = soup.find('b', text='Skin Type: ').next_sibling

skin_type

#### For all 

First

# skin_types_1 = []
# for l in links_1:
#     try:
#         response = requests.get(l)
#         soup = BeautifulSoup(response.content, 'html.parser')
#         skin_type = soup.find('b', text='Skin Type: ').next_sibling
#         print(skin_type)
#         skin_types_1.append(skin_type)
#     except:
#         skin_types_1.append('ERROR')

#skin_types_1= pd.DataFrame({'col':skin_types_1})

#skin_types_1.to_csv('../data/bs_skin_types_one.csv')

Second

# skin_types_2 = []
# for l in links_2:
#     try:
#         response = requests.get(l)
#         soup = BeautifulSoup(response.content, 'html.parser')
#         skin_type = soup.find('b', text='Skin Type: ').next_sibling
#         print(skin_type)
#         skin_types_2.append(skin_type)
#     except:
#         skin_types_2.append('ERROR')

#skin_types_2

#skin_types_2= pd.DataFrame({'col':skin_types_2})

#skin_types_2.to_csv('../data/bs_skin_types_two.csv')

Third

#links_3 = from_serie_to_list(data_3.URL)

#skin_types_3 = skin_type_category(links_3)

#skin_types_3.to_csv('../data/bs_skin_types_three.csv')

Fourth

#links_4 = from_serie_to_list(data_4.URL)

#skin_types_4 = skin_type_category(links_4)

#skin_types_4.to_csv('../data/bs_skin_types_four.csv')

Fifth

links_5 = from_serie_to_list(data_5.URL)

skin_types_5 = skin_type_category(links_5)

skin_types_5.to_csv('../data/bs_skin_types_five.csv')

Sixth

links_6 = from_serie_to_list(data_6.URL)

skin_types_6 = skin_type_category(links_6)

skin_types_6.to_csv('../data/bs_skin_types_six.csv')

Seventh

links_7 = from_serie_to_list(data_7.URL)

skin_types_7 = skin_type_category(links_7)

skin_types_7.to_csv('../data/bs_skin_types_seven.csv')

Eighth

links_8 = from_serie_to_list(data_8.URL)

skin_types_8 = skin_type_category(links_8)

skin_types_8.to_csv('../data/bs_skin_types_eight.csv')

Ninth

links_9 = from_serie_to_list(data_9.URL)

skin_types_9 = skin_type_category(links_9)

skin_types_9.to_csv('../data/bs_skin_types_nine.csv')
