==> notebooks/4-final_data.ipynb <==
import pandas as pd

import sys

 %load_ext autoreload
%autoreload 2

sys.path.append('../src')

from cleaning import concating_by_column, concating_by_row, new_index, rename_columns, order, droping_rows

### Final Dataframes

Joining all the dataset from scraping:

one = pd.read_csv('../data/bs_skin_types_one.csv', index_col='Unnamed: 0')

two = pd.read_csv('../data/bs_skin_types_two.csv', index_col='Unnamed: 0')

three = pd.read_csv('../data/bs_skin_types_three.csv', index_col='Unnamed: 0')

four = pd.read_csv('../data/bs_skin_types_four.csv', index_col='Unnamed: 0')

five = pd.read_csv('../data/bs_skin_types_five.csv', index_col='Unnamed: 0')

six = pd.read_csv('../data/bs_skin_types_six.csv', index_col='Unnamed: 0')

seven = pd.read_csv('../data/bs_skin_types_seven.csv', index_col='Unnamed: 0')

eight = pd.read_csv('../data/bs_skin_types_eight.csv', index_col='Unnamed: 0')

nine = pd.read_csv('../data/bs_skin_types_nine.csv', index_col='Unnamed: 0')

# Datasets to concat from scraping

to_concating = [one, two, three, four, five, six, seven, eight, nine]

# All datasets

final = concating_by_row(to_concating)

#  New index 

final = new_index(final)

# Export

final.to_csv('../data/skin_type_final.csv')

# Importing the final dataset

final_data = pd.read_csv('../data/skin_type_final_one.csv', index_col='Unnamed: 0')

# Looking for an ERROR

final_data[final_data.col == 'ERROR']

# Droping that row

final_data = droping_rows(final_data, [995])

# Looking for an ERROR ---> Not there anymore

final_data[final_data.col == 'ERROR']

# Regex to replace values

final_data =  final_data.replace( to_replace=r',.and', value= ',', regex=True)

# Regex to replace values

final_data = final_data.replace( to_replace=r'.and', value= ',', regex=True)

final_data.col  = final_data.col.str.strip()

skin_type_list = []
for d in final_data.col:
    skin_type_list.append(str(d).split(','))

list_as_df = pd.DataFrame({'col':skin_type_list})

list_as_df

# Export

final_data.to_csv('../data/final_data.csv')

### Final dataset

Joining the previous datasets with the original one

original = pd.read_csv('../data/clean_data.csv', index_col='Unnamed: 0')

# Dataset to concat

to_concat_final = [original, list_as_df]

final_dataset  =concating_by_column(to_concat_final)

# Rename columns

names = {'Brand':'Brand',
         'Category':'Category',
         'Name': 'Name',
         'Rating':'Rating',
         'Price':'Price',
         'URL':'URL',
         'Ingredients':'Ingredients',
         'col': 'Skin_Type'
}

final_dataset = rename_columns(final_dataset, names)

# New order for columns

columns = ['Brand', 'Category', 'Name', 'Skin_Type', 'Ingredients', 'URL', 'Price', 'Rating']

final_dataset = order(final_dataset, columns)

final_final_dataset = final_dataset.explode('Skin_Type')

# New index

final_final_dataset = new_index(final_final_dataset )

final_final_dataset.Skin_Type = final_final_dataset.Skin_Type.str.replace(" ", "")

# Drop na

final_final_dataset = final_final_dataset.dropna()

# Export

final_final_dataset.to_csv('../data/final_dataset_to_mongo.csv')
