==> notebooks/5-working_on_api.ipynb <==
import pandas as pd
from pymongo import MongoClient
import requests
import json

# API

data = pd.read_csv('../data/final_dataset_to_mongo.csv', index_col='Unnamed: 0')

data.sample(5)

## Mongo queries

For example:

con = MongoClient("localhost:27017")

db = con.get_database("skincare")

collection = db.get_collection("first")

# Single query

one = list(collection.find({'Brand':'Anthony'}, {'_id':0}))

one[0]

# Query using an input
# SKIN TYPE

user_input = input('Please choose a brand:')

two = list(collection.find({'Skin_Type':user_input}, {'_id':0}))

two[0]

len(two)

# Query using an multiple input
# SKIN TYPE

user_input = ['alpyn beauty', "Kiehl's Since 1851", 'Farmacy']

three = list(collection.find({'Brand':{"$in": user_input}}, {'_id':0}))

three[9]

# Multi query for routine

four = list(collection.find({"$and":[{'Skin_Type':'Oily'},{'Price':{"$lte":16.66}}, {'Category':'Cleanser'}]}, {'_id':0, 'Ingredients':0, 'URL':0}).limit(3))

four

five = list(collection.find({"$and":[{'Skin_Type':'Oily'},{'Price':{"$lte":16.66}}, {'Category':'Cleanser'}]}, {'_id':0, 'Brand':0, 'Category':0, 'Name':0, 'Skin_Type':0, 'Ingredients':0, 'Price':0, 'Rating':0}).limit(3).sort('Rating', -1))

five[0]

### Working on API options

data.sample(3)

list_brand = list(data.Brand.unique())

list_brand

list_category = list(data.Category.unique())

list_category

data.Skin_Type.unique()

mini = float(data.Price.min())

mini

data.Price.min()

maxi = int(data.Price.max())

maxi

data.Price.max()

data.Rating.max()

number = '98.65'

type(number)

number_number = float(number)

number_number

type(number_number)

data.Category.unique()

50/3

data[data.Category == 'Cleanser'].Price.max()

data[data.Category == 'Cleanser'].Price.min()

data[data.Category == 'Moisturizer'].Price.max()

data[data.Category == 'Moisturizer'].Price.min()

data[data.Category == 'Sunscreen & After Sun Care'].Price.max()

data[data.Category == 'Sunscreen & After Sun Care'].Price.min()
